feat: support download and conversion of other models
